+++ main.go
@@ -1,10 +1,9 @@
 package main
 
 import (
-	// "github.com/matthew1809/cart-tests/tests"
-	// "github.com/matthew1809/cart-tests/config"
-	// "net/http"
-	// "os"
+	"github.com/matthew1809/cart-tests/tests"
+	"github.com/matthew1809/cart-tests/config"
+	"net/http"
 	"flag"
 	"fmt"
 	"io/ioutil"
@@ -14,31 +13,32 @@ import (
 )
 
 func main() {
-	// variables := config.NewConfig()
-	// client := &http.Client{}
 
-	// tests.TestPromotions(variables, client)
-	// tests.RunTaxItemsTest(variables)
-	// tests.TestCheckout(variables, client)
-	// tests.TestCartWithTaxesAndPromotion(variables, client)
-
-	flag.String("test", "checkoutACart", "a valid test name")
+	test := flag.String("test", "", "a valid test name")
 	flag.Parse()
 
-	// fmt.Println("test:", *wordPtr)
+	if(*test == "") {
+		color.Cyan("Welcome to the API testing facility. You can run any of the following tests by running the program with a `-test=` flag followed by the test name:")
 
-	color.Cyan("Welcome to the API testing facility. You can run any of the following tests by running the program with a `-test=` flag followed by the test name:")
-	// switch argsWithoutProg := os.Args[1:]; argsWithoutProg {
-	// 	case
-	// }
+		files, err := ioutil.ReadDir("./tests")
+		if err != nil {
+			log.Fatal(err)
+		}
 
-	files, err := ioutil.ReadDir("./tests")
-    if err != nil {
-        log.Fatal(err)
-    }
+		for _, f := range files {
+			name := strings.Split(f.Name(), ".")
+			fmt.Println(name[0])
+		}
+	} else {
+		variables := config.NewConfig()
+		client := &http.Client{}
 
-    for _, f := range files {
-		name := strings.Split(f.Name(), ".")
-        fmt.Println(name[0])
-    }
+		switch *test {
+			case "TestPromotions": tests.TestPromotions(variables, client)
+			case "TestTaxItems": tests.TestTaxItems(variables, client)
+			case "TestCheckout": tests.TestCheckout(variables, client)
+			case "TestCartWithTaxesAndPromotion": tests.TestCartWithTaxesAndPromotion(variables, client)
+			default: fmt.Println("You have not entered a valid test")
+		}
+	}
 }
